pyproject.toml
[tool.poetry]
name = "mortgage"
version = "0.1.0"
description = ""
authors = ["Gareth Murphy <16783958+garethcmurphy@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pandas = "^2.2.3"
numpy = "^2.2.1"
matplotlib = "^3.10.0"
python-dotenv = "^1.0.1"
yfinance = "^0.2.51"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
